Split OS specific build files
