+++ subsystems/networking/networking_windows.go
@@ -1,0 +1,13 @@
+package networking
+
+import (
+	"context"
+
+	"github.com/viamrobotics/agent/subsystems"
+	"github.com/viamrobotics/agent/utils"
+	"go.viam.com/rdk/logging"
+)
+
+func NewSubsystem(_ context.Context, _ logging.Logger, _ utils.AgentConfig) subsystems.Subsystem {
+	return &subsystems.Dummy{}
+}
